docker_exploration/docker_realpython_CI_flask_tutorial/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
# created-by: conda 24.9.2
async-timeout=5.0.1=pyhd8ed1ab_0
blinker=1.8.2=pyhd8ed1ab_0
bzip2=1.0.8=h2466b09_7
ca-certificates=2024.8.30=h56e8100_0
click=8.1.7=win_pyh7428d3b_0
colorama=0.4.6=pyhd8ed1ab_0
coverage=7.6.4=py313hb4c8b1a_0
deprecated=1.2.14=pyh1a96a4e_0
exceptiongroup=1.2.2=pyhd8ed1ab_0
flask=3.0.3=py313haa95532_0
iniconfig=2.0.0=pyhd8ed1ab_0
itsdangerous=2.2.0=pyhd8ed1ab_0
jinja2=3.1.4=pyhd8ed1ab_0
libexpat=2.6.3=he0c23c2_0
libffi=3.4.2=h8ffe710_5
libmpdec=4.0.0=h2466b09_0
libsqlite=3.47.0=h2466b09_1
libzlib=1.3.1=h2466b09_2
markupsafe=3.0.2=py313hb4c8b1a_0
openssl=3.3.2=h2466b09_0
packaging=24.1=pyhd8ed1ab_0
pip=24.3.1=pyh145f28c_0
pluggy=1.5.0=pyhd8ed1ab_0
pytest=8.3.3=pyhd8ed1ab_0
python=3.13.0=hf5aa216_100_cp313
python_abi=3.13=5_cp313
redis-py=4.3.4=py313haa95532_0
setuptools=75.1.0=py313haa95532_0
tk=8.6.13=h5226925_1
tomli=2.0.2=pyhd8ed1ab_0
tzdata=2024b=hc8b5060_0
ucrt=10.0.22621.0=h57928b3_1
vc=14.3=h8a93ad2_22
vc14_runtime=14.40.33810=hcc2c482_22
vs2015_runtime=14.40.33810=h3bf8584_22
werkzeug=3.1.0=pyhd8ed1ab_0
wrapt=1.16.0=py313ha7868ed_1
xz=5.2.6=h8d14728_0
